NetworkManager

+++ ConsoleAdventure/Content/Scripts/Player/PlayerInfo.cs
@@ -2,6 +2,6 @@ namespace ConsoleAdventure.Content.Scripts.Player;
 
 public class PlayerInfo
 {
-    public int Id;
+    public short Id;
     public string Name = "William";
 }
